File upload download

+++ controllers/admin.js
@@ -1,9 +1,11 @@
-const { error } = require('console');
+
 const mongoDB = require('mongodb')
 const Product = require('../models/product');
 const product = require('../models/product');
 const {validationResult} = require('express-validator');
 const { default: mongoose } = require('mongoose');
+const fileHelper = require('../util/file')
+
 exports.getAddProduct = (req, res, next) => {
 
   res.render('admin/edit-product', {
@@ -19,17 +21,34 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  let image= req.file
   const price = req.body.price;
   const description = req.body.description;
-const errors = validationResult(req)
+  const errors = validationResult(req)
+  
+  if(!req?.file){
+    console.log("here file", req?.file)
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      product: {
+        title, price, description
+      },
+      hasError : true,
+      isAuthenticated : req.session.isLoggedIn,
+      errorMessage : 'Attached file is not an image',
+      validationErrors: []
+    })
+  } 
+
 if(!errors.isEmpty()){
  return res.status(422).render('admin/edit-product', {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
     editing: false,
     product: {
-      title, imageUrl, price, description
+      title, image, price, description
     },
     hasError : true,
     isAuthenticated : req.session.isLoggedIn,
@@ -37,6 +56,8 @@ if(!errors.isEmpty()){
     validationErrors: errors.array()
   })
 }
+
+const imageUrl = image.path
   const product = new Product({
     title,
     price,
@@ -54,6 +75,7 @@ if(!errors.isEmpty()){
       }
     )
     .catch(err=> {
+      console.log("here error", err)
       const error = new Error(err)
       error.httpStatusCode = 500
       return next(error)
@@ -98,8 +120,8 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId
   const updatedTitle = req.body.title
   const updatedPrice = req.body.price
-  const updatedImageUrl = req.body.imageUrl
   const updatedDescription = req.body.description
+  const image = req.file
   const errors = validationResult(req)
   if(!errors.isEmpty()){
    return res.status(422).render('admin/edit-product', {
@@ -107,8 +129,7 @@ exports.postEditProduct = (req, res, next) => {
       path: '/admin/edit-product',
       editing: true,
       product: {
-        title : updatedTitle,
-         imageUrl:  updatedImageUrl, price: updatedPrice, description : updatedDescription,
+        title : updatedTitle,price: updatedPrice, description : updatedDescription,
          _id : prodId
       },
       hasError : true,
@@ -127,7 +148,11 @@ exports.postEditProduct = (req, res, next) => {
       result.title = updatedTitle
       result.price = updatedPrice
       result.description = updatedDescription
-      result.imageUrl = updatedImageUrl
+      if(image){
+        fileHelper.deleteFile(product.imageUrl)
+        //if new file is uploaded then save it. otherwise dont
+        result.imageUrl = image.path
+      }
 
       return result.save().then(result => {
         console.log("updated product")
@@ -162,16 +187,24 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId
-  Product.deleteOne({_id : productId, userId : req.user._id})
+Product.findById(productId).then(product => {
+if(!product){
+  return next(new Error('Product not found'))
+}
 
-    .then(result => {
-      console.log('deleted product')
-      res.redirect('/admin/products')
+fileHelper.deleteFile(product.imageUrl)
+return Product.deleteOne({_id : productId, userId : req.user._id})
 
-    })
-    .catch(err=> {
-      const error = new Error(err)
-      error.httpStatusCode = 500
-      return next(error)
-    })
+.then(result => {
+  console.log('deleted product')
+  res.redirect('/admin/products')
+
+})
+.catch(err=> {
+  const error = new Error(err)
+  error.httpStatusCode = 500
+  return next(error)
+})
+})
+  
 }
